Fixed arrow alignment in FAQ
